Teach bb-editor about topGraphLog
