+++ packages/shared-ui/src/elements/editor/editor.ts
@@ -50,6 +50,7 @@ const DATA_TYPE = "text/plain";
 const PASTE_OFFSET = 50;
 
 import { GraphAssets } from "./graph-assets.js";
+import { LogEntry, NodeLogEntry } from "../../types/types.js";
 
 function getDefaultConfiguration(type: string): NodeConfiguration | undefined {
   if (type !== "input" && type !== "output") {
@@ -103,7 +104,7 @@ export class Editor extends LitElement {
   mode = EditorMode.ADVANCED;
 
   @property()
-  highlightedNodeId: string | null = null;
+  topGraphLog: LogEntry[] | null = null;
 
   @state()
   nodeValueBeingEdited: EditedNode | null = null;
@@ -1260,8 +1261,24 @@ export class Editor extends LitElement {
     this.#onResizeBound();
   }
 
+  #currentNode() {
+    if (!this.topGraphLog) return null;
+
+    // @ts-expect-error -- TS doesn't know findLastIndex exists
+    const currentNode = this.topGraphLog.findLast((entry) => {
+      return entry.type === "node";
+    }) as NodeLogEntry | undefined;
+
+    if (!currentNode) return null;
+
+    if (this.subGraphId) return null;
+
+    return currentNode;
+  }
+
   render() {
-    this.#graphRenderer.highlightedNodeId = this.highlightedNodeId;
+    this.#graphRenderer.highlightedNodeId =
+      this.#currentNode()?.descriptor.id || null;
 
     if (this.#graphRenderer) {
       this.#graphRenderer.invertZoomScrollDirection =
